--- test/helpers/foundry/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "allowSyntheticDefaultImports": true,
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "target": "ESNext"
-  },
-  "extends": "../../../tsconfig.json"
-}
